Added Layer class and member function implementations

+++ common/include/isotope_data.h
@@ -45,11 +45,9 @@ namespace common {
  */
 class IsotopeData {
  public:
-  static constexpr unsigned int MAX_Z =
-      118;  //!< The maximum allowed atomic number
+  static constexpr size_t MAX_Z = 118;  //!< The maximum allowed atomic number
 
-  static constexpr unsigned int MAX_A =
-      295;  //!< The maximum allowed mass number
+  static constexpr size_t MAX_A = 295;  //!< The maximum allowed mass number
 
   /*!
    * \brief A function that takes a Z number and mass number, and returns the
@@ -62,8 +60,8 @@ class IsotopeData {
    * \param mass_number The mass number of the isotope
    * \return The exact mass of the corresponding isotope in Relative Atomic Mass
    */
-  static constexpr mass_quantity getIsotopeMass(unsigned int atomic_number,
-                                                unsigned int mass_number) {
+  static constexpr mass_quantity getIsotopeMass(size_t atomic_number,
+                                                size_t mass_number) {
     if (atomic_number < 1 || atomic_number > MAX_Z) {
       throw std::out_of_range(
           "Invalid atomic number: " + std::to_string(atomic_number) +
